refactor: Rname Expenses naming with Transactions across components and routes
- Deleted the ExpensesPage component and replaced it with TransactionsPage.
- Updated MoniteComponents to export Transactions instead of Expenses.
- Modified NavigationList and App sidebar to link to Transactions instead of Expenses.
- Introduced TransactionsTable component with necessary hooks and filters.
- Updated component settings to reflect the change from expenses to transactions.
- Removed obsolete ExpensesTable and related files.
