--- examples/with-nextjs-and-clerk-auth/src/app/(monite)/expenses/page.tsx
@@ -1,10 +1,0 @@
-import { Expenses } from '@/components/MoniteComponents';
-import { Box } from '@mui/material';
-
-export default async function ExpensesPage() {
-  return (
-    <Box className="Monite-PageContainer Monite-Expenses">
-      <Expenses />
-    </Box>
-  );
-}
