+++ examples/with-nextjs-and-clerk-auth/src/app/(monite)/transactions/page.tsx
@@ -1,0 +1,10 @@
+import { Transactions } from '@/components/MoniteComponents';
+import { Box } from '@mui/material';
+
+export default async function TransactionsPage() {
+  return (
+    <Box className="Monite-PageContainer Monite-Transactions">
+      <Transactions />
+    </Box>
+  );
+}
