+++ examples/with-nextjs-and-clerk-auth/src/components/NavigationMenu/NavigationList.tsx
@@ -35,8 +35,8 @@ export const NavigationList = () => {
         <NavigationListItem href="/receivables" icon={<IconReceipt />}>
           {t(i18n)`Invoicing`}
         </NavigationListItem>
-        <NavigationListItem href="/expenses" icon={<IconUsdCircle />}>
-          {t(i18n)`Expenses`}
+        <NavigationListItem href="/transactions" icon={<IconUsdCircle />}>
+          {t(i18n)`Transactions`}
         </NavigationListItem>
         <NavigationListItem href="/projects" icon={<IconBag />}>
           {t(i18n)`Projects`}
