--- packages/sdk-react/src/components/expenses/ExpensesTable/index.ts
@@ -1,2 +1,0 @@
-export { UserTransactionsTable } from './UserTransactionsTable';
-export { ManagerTransactionsTable } from './ManagerTransactionsTable';
