+++ packages/sdk-react/src/components/expenses/TransactionsTable/TransactionsTable.tsx
@@ -1,30 +1,22 @@
+import { ReceiptPreview } from '../ReceiptPreview';
 import { TransactionDetails } from '../TransactionDetails';
 import { useGetTransactions } from '../hooks/useGetTransactions';
-import {
-  FILTER_TYPE_SEARCH,
-  FILTER_TYPE_STARTED_AT,
-  FILTER_TYPE_USER,
-} from './consts';
+import { FILTER_TYPE_SEARCH, FILTER_TYPE_STARTED_AT } from './consts';
 import type { FilterTypes } from './types';
-import { components } from '@/api';
-import { UserDisplayCell } from '@/components/UserDisplayCell/UserDisplayCell';
+import { type components } from '@/api';
 import { useMoniteContext } from '@/core/context/MoniteContext';
 import { useRootElements } from '@/core/context/RootElementsProvider';
 import { useCurrencies, useDataTableState } from '@/core/hooks';
 import { useDebounce } from '@/core/hooks/useDebounce';
-import { useEntityUserByAuthToken, useEntityUsersByIds } from '@/core/queries';
+import { useEntityUserByAuthToken } from '@/core/queries';
 import { useIsActionAllowed } from '@/core/queries/usePermissions';
+import { useReceiptsByTransactionIds } from '@/core/queries/useReceiptsByTransactionIds';
 import { GetNoRowsOverlay } from '@/ui/DataGridEmptyState/GetNoRowsOverlay';
 import { AccessRestriction } from '@/ui/accessRestriction';
+import { Button } from '@/ui/components/button';
 import { DataTable } from '@/ui/components/data-table';
 import { Input } from '@/ui/components/input';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/ui/components/select';
+import { Skeleton } from '@/ui/components/skeleton';
 import { LoadingPage } from '@/ui/loadingPage';
 import { hasSelectedText } from '@/utils/text-selection';
 import { t } from '@lingui/macro';
@@ -34,7 +26,7 @@ import { ColumnDef } from '@tanstack/react-table';
 import { formatISO, addDays } from 'date-fns';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
-export const ManagerTransactionsTable = () => {
+export const TransactionsTable = () => {
   const { componentSettings, locale } = useMoniteContext();
   const { i18n } = useLingui();
   const { root } = useRootElements();
@@ -44,22 +36,15 @@ export const ManagerTransactionsTable = () => {
     components['schemas']['TransactionResponse'] | undefined
   >(undefined);
   const [detailsModalOpened, setDetailsModalOpened] = useState<boolean>(false);
+  const [receiptPreviewReceipt, setReceiptPreviewReceipt] = useState<
+    components['schemas']['ReceiptResponseSchema'] | undefined
+  >(undefined);
 
   const { data: user } = useEntityUserByAuthToken();
 
-  const {
-    data: isTransactionReadSupported,
-    isLoading: isTransaxtionReadSupportedLoading,
-  } = useIsActionAllowed({
-    method: 'transaction',
-    action: 'read',
-    entityUserId: user?.id,
-  });
-
-  // TODO: refactor to use proper check of Expense Management manager
-  const { data: isUserReadSupported, isLoading: isUserReadSupportedLoading } =
+  const { data: isReadSupported, isLoading: isReadSupportedLoading } =
     useIsActionAllowed({
-      method: 'entity_user',
+      method: 'transaction',
       action: 'read',
       entityUserId: user?.id,
     });
@@ -79,7 +64,7 @@ export const ManagerTransactionsTable = () => {
     components['schemas']['TransactionCursorFields'],
     FilterTypes
   >({
-    initialPageSize: componentSettings.expenses.pageSizeOptions[0] ?? 20,
+    initialPageSize: componentSettings.transactions.pageSizeOptions[0] ?? 20,
     initialSort: {
       field: 'started_at',
       direction: 'desc',
@@ -88,7 +73,6 @@ export const ManagerTransactionsTable = () => {
       initialValue: {
         [FILTER_TYPE_SEARCH]: '',
         [FILTER_TYPE_STARTED_AT]: null,
-        [FILTER_TYPE_USER]: null,
       },
     },
   });
@@ -129,6 +113,7 @@ export const ManagerTransactionsTable = () => {
         filters[FILTER_TYPE_SEARCH] && pagination.pageIndex === 0
           ? undefined
           : currentPaginationToken || undefined,
+      entity_user_id__in: user?.id ? [user.id] : undefined,
       merchant_name__icontains: filters[FILTER_TYPE_SEARCH] || undefined,
       started_at__gt: filters[FILTER_TYPE_STARTED_AT]
         ? formatISO(filters[FILTER_TYPE_STARTED_AT] as Date)
@@ -136,11 +121,8 @@ export const ManagerTransactionsTable = () => {
       started_at__lt: filters[FILTER_TYPE_STARTED_AT]
         ? formatISO(addDays(filters[FILTER_TYPE_STARTED_AT] as Date, 1))
         : undefined,
-      entity_user_id__in: filters[FILTER_TYPE_USER]
-        ? [filters[FILTER_TYPE_USER]]
-        : undefined,
     },
-    isTransactionReadSupported === true && isUserReadSupported === true
+    isReadSupported === true
   );
 
   // Update the pagination hook with the latest API response
@@ -148,73 +130,19 @@ export const ManagerTransactionsTable = () => {
     updateApiResponse(transactionsResponse);
   }, [transactionsResponse, updateApiResponse]);
 
-  // Extract unique user IDs from transactions data
-  const uniqueUserIds = useMemo(() => {
+  // Extract transaction IDs for receipts query
+  const transactionIds = useMemo(() => {
     if (!transactions) return [];
-
-    const userIds = new Set<string>();
-
-    transactions.forEach((transaction) => {
-      if (transaction.entity_user_id) {
-        userIds.add(transaction.entity_user_id);
-      }
-    });
-
-    return Array.from(userIds);
+    return transactions.map((transaction) => transaction.id);
   }, [transactions]);
 
-  // Fetch user details for the unique user IDs
-  const { data: usersData } = useEntityUsersByIds(uniqueUserIds);
-
-  // Create user data map and filter array
-  const { userDataMap, users } = useMemo(() => {
-    if (!usersData?.data) return { userDataMap: new Map(), users: [] };
-
-    const userMap = new Map();
-    const usersArray = usersData.data.map((user) => {
-      userMap.set(user.id, user);
-      return {
-        id: user.id,
-        name:
-          `${user.first_name ?? ''} ${user.last_name ?? ''}`.trim() ||
-          user.email ||
-          // eslint-disable-next-line lingui/no-unlocalized-strings
-          `User ${user.id.slice(0, 8)}...`,
-      };
-    });
-
-    return { userDataMap: userMap, users: usersArray };
-  }, [usersData?.data]);
-
-  const UserCell = useMemo(() => {
-    return ({ userId }: { userId: string | null }) => {
-      if (!userId) return <span>-</span>;
-
-      const userData = userDataMap.get(userId);
-      if (!userData) return <span>-</span>;
-
-      return (
-        <UserDisplayCell
-          user={userData}
-          showAvatar={false}
-          variant="inline"
-          typographyVariant="inherit"
-        />
-      );
-    };
-  }, [userDataMap]);
+  // Fetch receipts for the transaction IDs
+  const { receiptsByTransactionId, isLoading: isReceiptsLoading } =
+    useReceiptsByTransactionIds(transactionIds);
 
   const columns: ColumnDef<components['schemas']['TransactionResponse']>[] =
     useMemo(
       () => [
-        {
-          header: t(i18n)`Employee`,
-          accessorKey: 'employee',
-          cell: ({ row }) => (
-            <UserCell userId={row.original.entity_user_id ?? null} />
-          ),
-          enableSorting: false,
-        },
         {
           header: t(i18n)`Merchant`,
           accessorKey: 'merchant_name',
@@ -257,6 +185,48 @@ export const ManagerTransactionsTable = () => {
           enableSorting: false,
         },
         {
+          header: t(i18n)`Receipt`,
+          id: 'receipt',
+          accessorFn: (row) => {
+            const receipt = receiptsByTransactionId[row.id];
+            return receipt ? receipt.file_url || receipt.file_id : null;
+          },
+          cell: ({ row }) => {
+            const receipt = receiptsByTransactionId[row.original.id];
+
+            if (isReceiptsLoading) {
+              return <Skeleton className="mtw:w-full mtw:h-4" />;
+            }
+
+            if (!receipt) {
+              return (
+                <span className="mtw:text-muted-foreground">{t(
+                  i18n
+                )`Not matched`}</span>
+              );
+            }
+
+            if (receipt.file_url) {
+              return (
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  onClick={(e) => {
+                    e.stopPropagation();
+                    setReceiptPreviewReceipt(receipt);
+                  }}
+                  className="mtw:text-primary mtw:font-medium mtw:text-sm mtw:p-0 mtw:h-5"
+                >
+                  {t(i18n)`Receipt ${receipt.document_id}`}
+                </Button>
+              );
+            }
+
+            return <span>{t(i18n)`Receipt ${receipt.document_id}`}</span>;
+          },
+          enableSorting: false,
+        },
+        {
           header: t(i18n)`Amount`,
           accessorKey: 'amount',
           cell: ({ row }) => {
@@ -269,7 +239,13 @@ export const ManagerTransactionsTable = () => {
           },
         },
       ],
-      [UserCell, formatCurrencyToDisplay, i18n, locale.dateTimeFormat]
+      [
+        i18n,
+        locale.dateTimeFormat,
+        receiptsByTransactionId,
+        isReceiptsLoading,
+        formatCurrencyToDisplay,
+      ]
     );
 
   const openDetailsModal = useCallback(
@@ -280,6 +256,7 @@ export const ManagerTransactionsTable = () => {
 
       // Close any existing modal and receipt preview
       setDetailsModalOpened(false);
+      setReceiptPreviewReceipt(undefined);
 
       // Use setTimeout to ensure clean state transition
       setTimeout(() => {
@@ -295,11 +272,11 @@ export const ManagerTransactionsTable = () => {
     setDetailsModalOpened(false);
   }, []);
 
-  if (isTransaxtionReadSupportedLoading || isUserReadSupportedLoading) {
+  if (isReadSupportedLoading) {
     return <LoadingPage />;
   }
 
-  if (!isTransactionReadSupported || !isUserReadSupported) {
+  if (!isReadSupported) {
     return <AccessRestriction />;
   }
 
@@ -312,25 +289,7 @@ export const ManagerTransactionsTable = () => {
           onChange={(event) => setSearchInputValue(event.target.value)}
           className="mtw:max-w-sm"
         />
-        <div className="mtw:flex mtw:items-center mtw:gap-4">
-          <Select
-            value={filters[FILTER_TYPE_USER] || 'all'}
-            onValueChange={(value) =>
-              onFilterChange(FILTER_TYPE_USER, value === 'all' ? null : value)
-            }
-          >
-            <SelectTrigger className="mtw:w-[200px]">
-              <SelectValue placeholder={t(i18n)`Filter by employee`} />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="all">{t(i18n)`All employees`}</SelectItem>
-              {users.map((user) => (
-                <SelectItem key={user.id} value={user.id}>
-                  {user.name}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
+        <div className="mtw:flex mtw:items-center mtw:shrink mtw:gap-2 mtw:justify-between">
           <DatePicker
             className="Monite-ExpensesStartedAtFilter Monite-FilterControl Monite-DateFilterControl"
             onChange={(value, error) => {
@@ -386,17 +345,15 @@ export const ManagerTransactionsTable = () => {
             <GetNoRowsOverlay
               isLoading={isLoading}
               dataLength={transactions?.length || 0}
-              isFiltering={
-                !!filters[FILTER_TYPE_STARTED_AT] || !!filters[FILTER_TYPE_USER]
-              }
+              isFiltering={!!filters[FILTER_TYPE_STARTED_AT]}
               isSearching={!!filters[FILTER_TYPE_SEARCH]}
               isError={!!error}
               refetch={refetch}
-              entityName={t(i18n)`Expenses`}
-              noDataTitle={t(i18n)`No expenses yet`}
-              noDataDescription1={t(i18n)`You don’t have any expenses yet`}
-              noDataDescription2={t(i18n)`Your expenses will appear here`}
-              filterTitle={t(i18n)`No expenses found`}
+              entityName={t(i18n)`Transactions`}
+              noDataTitle={t(i18n)`No transactions yet`}
+              noDataDescription1={t(i18n)`You don't have any transactions yet`}
+              noDataDescription2={t(i18n)`Your transactions will appear here`}
+              filterTitle={t(i18n)`No transactions found`}
               filterDescription1={t(
                 i18n
               )`Try adjusting your search or filter criteria`}
@@ -406,11 +363,21 @@ export const ManagerTransactionsTable = () => {
         />
       </div>
       <TransactionDetails
-        isManagerView={true}
         transaction={selectedTransaction}
         open={detailsModalOpened}
         onClose={closeDetailsModal}
       />
+      {receiptPreviewReceipt && (
+        <ReceiptPreview
+          receipt={receiptPreviewReceipt}
+          isOpen={!!receiptPreviewReceipt}
+          setIsOpen={(isOpen) => {
+            if (!isOpen) {
+              setReceiptPreviewReceipt(undefined);
+            }
+          }}
+        />
+      )}
     </div>
   );
 };
