+++ packages/sdk-react/src/components/expenses/TransactionsTable/consts.ts
