+++ packages/sdk-react/src/components/expenses/TransactionsTable/index.ts
@@ -1,0 +1,1 @@
+export { TransactionsTable } from './TransactionsTable';
